Extract broadcastToArea func.
